controller

+++ controllers/contactController.js
@@ -51,6 +51,11 @@ const updateContact = asyncHanlder(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("Use don't have permission to update other user contacts");
+  }
+
   const updateContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -69,6 +74,12 @@ const deleteContact = asyncHanlder(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("Use don't have permission to delete other user contacts");
+  }
+
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
 });
